refactor(processor): extract page result assembly
Move PageResult projection, initial quality classification, and native search matching into a focused processor helper. This keeps processDocument centered on pipeline orchestration while leaving the public page payload unchanged.
